Refactor parfor cache tests to make use of existing code.
As title.
