+++ numba/tests/test_parfors_caching.py
@@ -1,4 +1,6 @@
 import os.path
+import subprocess
+import sys
 
 import numpy as np
 
@@ -7,7 +9,7 @@ from .test_caching import DispatcherCacheUsecasesTest
 
 
 @skip_parfors_unsupported
-class TestParForsCache(DispatcherCacheUsecasesTest):
+class TestParforsCache(DispatcherCacheUsecasesTest):
     here = os.path.dirname(__file__)
     usecases_file = os.path.join(here, "parfors_cache_usecases.py")
     modname = "parfors_caching_test_fodder"
@@ -41,3 +43,39 @@ class TestParForsCache(DispatcherCacheUsecasesTest):
         # num_funcs=3 because, there's the `caller_case()` which calls
         # the `prange_case()` and `arrayexprs_case()`
         self.run_test(f, num_funcs=3)
+
+
+@skip_parfors_unsupported
+class TestParforsCacheChangingThreads(DispatcherCacheUsecasesTest):
+    # NOTE: This test is checking issue #7518, that thread counts are not
+    # baked into cached objects.
+
+    here = os.path.dirname(__file__)
+    usecases_file = os.path.join(here, "parfors_cache_usecases.py")
+    modname = "parfors_caching_test_fodder"
+
+    def run_in_separate_process(self, thread_count):
+        # Cached functions can be run from a distinct process.
+        code = """if 1:
+            import sys
+
+            sys.path.insert(0, %(tempdir)r)
+            mod = __import__(%(modname)r)
+            mod.self_run()
+            """ % dict(tempdir=self.tempdir, modname=self.modname)
+
+        new_env = {**os.environ, "NUMBA_NUM_THREADS" : str(thread_count)}
+        popen = subprocess.Popen([sys.executable, "-c", code],
+                                 stdout=subprocess.PIPE, stderr=subprocess.PIPE,
+                                 env=new_env)
+        out, err = popen.communicate()
+        if popen.returncode != 0:
+            raise AssertionError(f"process failed with code {popen.returncode}:"
+                                 f"stderr follows\n{err.decode()}\n")
+
+    def test_caching(self):
+        self.check_pycache(0)
+        self.run_in_separate_process(1)
+        self.check_pycache(3 * 2) # ran 3 functions, 2 entries each
+        self.run_in_separate_process(2)
+        self.check_pycache(3 * 2) # ran 3 functions, 2 entries each
